+++ SyntaxHighlighting.c
@@ -17,37 +17,28 @@
 #include <windows.h>
 #include <stdlib.h>
 
+#include "grammar.h"
 #include "syntaxhighlighting.h"
 #include "editorfont.h"
 
-#define MAX_TOKEN_TYPE	64
 #define WINDOW_SIZE		100
 
 typedef unsigned char* (*HIGHLIGHT_CALCULATE)(HIGHLIGHTER* pData, FTABLE* fp, LINE* lp, long nLine);
 
-typedef enum {
-	UNKNOWN,
-	OTHER,
-	STRING,
-	IDENTIFIER,
-	OPERATOR,
-	KEYWORD,
-	SL_COMMENT,
-	ML_COMMENT
-} TOKEN_TYPE;
-
 typedef struct tagTOKEN_LINE_CACHE {
 	long s_lineNumber;
 	LINE* s_linePointer;
-	TOKEN_TYPE s_lineEndType;
+	LEXICAL_STATE s_lineEndType;
 } TOKEN_LINE_CACHE;
 
 typedef struct tagHIGHLIGHTER {
 	HIGHLIGHT_CALCULATE h_calculate;
 	unsigned char h_tokenTypeToStyleTable[MAX_TOKEN_TYPE];
 	unsigned char* h_styles;
+	long h_styleCapacity;
 	long h_minLine;
 	long h_lastLine;
+	GRAMMAR* h_grammar;
 	LINE* h_lastLinePointer;
 	TOKEN_LINE_CACHE h_lines[WINDOW_SIZE];
 } HIGHLIGHTER;
@@ -58,11 +49,9 @@ static void highlight_destroyCaches(HIGHLIGHTER* pHighlighter) {
 	if (pHighlighter->h_styles) {
 		free(pHighlighter->h_styles);
 		pHighlighter->h_styles = 0;
+		pHighlighter->h_styleCapacity = 0;
 	}
-	for (int i = 0; i < WINDOW_SIZE; i++) {
-		pHighlighter->h_lines[i].s_linePointer = 0;
-		pHighlighter->h_lines[i].s_lineEndType = UNKNOWN;
-	}
+	memset(pHighlighter->h_lines, 0, sizeof(pHighlighter->h_lines));
 }
 
 /*
@@ -73,104 +62,110 @@ void highlight_destroy(HIGHLIGHTER* pHighlighter) {
 	free(pHighlighter);
 }
 
-static TOKEN_TYPE highlight_getPreviousLineTokenType(HIGHLIGHTER* pHighlighter, FTABLE* fp, LINE* lp, long nLine) {
+static void highlight_adjustCachedLineWindow(HIGHLIGHTER* pHighlighter, FTABLE* fp, long nLine) {
+	int nDelta = nLine - pHighlighter->h_minLine;
+	if (nDelta < 10 || nDelta >= WINDOW_SIZE-10) {
+		nLine -= WINDOW_SIZE / 2;
+		if (nLine >= fp->nlines) {
+			nLine = fp->nlines - WINDOW_SIZE;
+		}
+		if (nLine < 0) {
+			nLine = 0;
+		}
+		int nDelta = nLine- pHighlighter->h_minLine;
+		if (nDelta <= -WINDOW_SIZE || nDelta >= WINDOW_SIZE) {
+			highlight_destroyCaches(pHighlighter);
+		}
+		else if (nDelta < 0) {
+			memmove(&pHighlighter->h_lines[-nDelta], pHighlighter->h_lines, (WINDOW_SIZE + nDelta) * sizeof(TOKEN_LINE_CACHE));
+			memset(pHighlighter->h_lines, 0, -nDelta * sizeof(TOKEN_LINE_CACHE));
+		}
+		else if (nDelta > 0) {
+			memmove(pHighlighter->h_lines, &pHighlighter->h_lines[nDelta], (WINDOW_SIZE - nDelta) * sizeof(TOKEN_LINE_CACHE));
+			memset(&pHighlighter->h_lines[WINDOW_SIZE - nDelta], 0, nDelta * sizeof(TOKEN_LINE_CACHE));
+		}
+		pHighlighter->h_minLine = nLine;
+	}
+}
+
+static LEXICAL_STATE highlight_getPreviousLineTokenType(HIGHLIGHTER* pHighlighter, FTABLE* fp, LINE* lp, long nLine) {
 	if (nLine == 0) {
-		return OTHER;
+		return INITIAL;
 	}
 	int nPreviousIdx = nLine - pHighlighter->h_minLine - 1;
 	if (nPreviousIdx < 0 || nPreviousIdx >= WINDOW_SIZE) {
-		highlight_destroyCaches(pHighlighter);
-		for (int i = 0; i < WINDOW_SIZE / 2 && lp && nLine >= 0; i++) {
-			nLine--;
-			lp = lp->prev;
-		}
-		pHighlighter->h_minLine = nLine;
-		// TODO: determine line-spanning tokens from previous lines, when the cache is invalidated.
-		return OTHER;
+		return UNKNOWN;
 	}
-	TOKEN_TYPE tType = pHighlighter->h_lines[nPreviousIdx].s_lineEndType;
-	return tType == UNKNOWN ? OTHER : tType;
+	LINE* lpPrev = lp->prev;
+	TOKEN_LINE_CACHE* tlp = &pHighlighter->h_lines[nPreviousIdx];
+	if (tlp->s_linePointer != lpPrev) {
+		// TODO: determine line-spanning tokens from previous lines, when the cache is invalidated.
+		return UNKNOWN;
+	}
+	LEXICAL_STATE tType = tlp->s_lineEndType;
+	if (tType == UNKNOWN) {
+		return INITIAL;
+	}
+	return tType;
 }
 
 /*
- * A simple demo highlighter.
+ * A highlighter which passes on tokenization to a generic grammar driven tokenizer.
  */
-static unsigned char* highlight_demo(HIGHLIGHTER* pHighlighter, FTABLE* fp, LINE* lp, long nLine) {
-	TOKEN_TYPE tokenType = highlight_getPreviousLineTokenType(pHighlighter, fp, lp, nLine);
-	if (tokenType == SL_COMMENT) {
-		tokenType = OTHER;
-	}
+static unsigned char* highlight_usingGrammar(HIGHLIGHTER* pHighlighter, FTABLE* fp, LINE* lp, long nLine) {
+	highlight_adjustCachedLineWindow(pHighlighter, fp, nLine);
+	LEXICAL_STATE lexicalState = highlight_getPreviousLineTokenType(pHighlighter, fp, lp, nLine);
 	if (nLine != pHighlighter->h_lastLine || lp != pHighlighter->h_lastLinePointer) {
-		pHighlighter->h_styles = pHighlighter->h_styles ? realloc(pHighlighter->h_styles, lp->len) : malloc(lp->len);
+		if (pHighlighter->h_styles) {
+			if (lp->len >= pHighlighter->h_styleCapacity - 10) {
+				pHighlighter->h_styleCapacity = lp->len+30;
+				pHighlighter->h_styles = realloc(pHighlighter->h_styles, pHighlighter->h_styleCapacity);
+			}
+		} else {
+			pHighlighter->h_styleCapacity = lp->len+30;
+			pHighlighter->h_styles = malloc(pHighlighter->h_styleCapacity);
+		}
 		unsigned char* pStyles = pHighlighter->h_styles;
 		pHighlighter->h_lastLinePointer = lp;
 		pHighlighter->h_lastLine = nLine;
-		for (int i = 0; i < lp->len; i++) {
-			unsigned char c = lp->lbuf[i];
-			if (c <= ' ') {
-				*pStyles++ = FS_CONTROL_CHARS;
-				continue;
-			} else if (tokenType == ML_COMMENT) {
-				if (c == '*' && i + 1 < lp->len && lp->lbuf[i + 1] == '/') {
-					i++;
-					tokenType = OTHER;
-					*pStyles++ = FS_COMMENT;
-					*pStyles++ = FS_COMMENT;
-					continue;
-				}
-			} else if (tokenType == STRING) {
-				if (c == '\\' && i + 1 < lp->len) {
-					i++;
-					*pStyles++ = FS_STRING;
-					*pStyles++ = FS_STRING;
-					continue;
-				}
-				else if (c == '"') {
-					*pStyles++ = FS_STRING;
-					tokenType = OTHER;
-					continue;
-				}
-			} else if (tokenType != SL_COMMENT) {
-				TOKEN_TYPE hOld = tokenType;
-				if (c == '/' && i + 1 < lp->len) {
-					if (lp->lbuf[i + 1] == '/') {
-						tokenType = SL_COMMENT;
-					} else if (lp->lbuf[i + 1] == '*') {
-						tokenType = ML_COMMENT;
-					}
-					if (tokenType != hOld) {
-						i++;
-						*pStyles++ = FS_COMMENT;
-						*pStyles++ = FS_COMMENT;
-						continue;
-					}
-				} else if (c == '"') {
-					tokenType = STRING;
-				} else if (c == '*' || c == '?' || c == '/' || c == '~' || c == '-' || c == '=' || c == '!' || c == '|' || c == '&') {
-					tokenType = OPERATOR;
-				} else {
-					tokenType = OTHER;
-				}
+		LEXICAL_ELEMENT lexicalElements[MAX_LEXICAL_ELEMENT];
+		int nElements = grammar_parse(pHighlighter->h_grammar, lexicalElements, lexicalState, lp->lbuf, lp->len);
+		int nPreviousOffset = 0;
+		if (nElements == 0) {
+			for (int i = 0; i < lp->len; i++) {
+				unsigned char c = lp->lbuf[i];
+				pStyles[i] = (c <= ' ') ? FS_CONTROL_CHARS : FS_NORMAL;
 			}
-			*pStyles++ = pHighlighter->h_tokenTypeToStyleTable[tokenType];
+		} else {
+			for (int i = 0; i < nElements; i++) {
+				int nNextOffset = nPreviousOffset + lexicalElements[i].le_length;
+				lexicalState = lexicalElements[i].le_state;
+				for (int j = nPreviousOffset; j < nNextOffset && j < lp->len; j++) {
+					unsigned char c = lp->lbuf[j];
+					pStyles[j] = (c <= ' ') ? FS_CONTROL_CHARS : pHighlighter->h_tokenTypeToStyleTable[lexicalState];
+				}
+				nPreviousOffset = nNextOffset;
+			}
 		}
-		pHighlighter->h_lines[nLine - pHighlighter->h_minLine].s_lineEndType = tokenType;
+		int nOffset = nLine - pHighlighter->h_minLine;
+		if (nOffset >= 0 && nOffset < WINDOW_SIZE) {
+			TOKEN_LINE_CACHE* tlp = &pHighlighter->h_lines[nOffset];
+			tlp->s_lineEndType = lexicalState;
+			tlp->s_linePointer = lp;
+			tlp->s_lineNumber = nLine;
+		}
 	}
 	return pHighlighter->h_styles;
 }
-
+	
 /*
  * Invalidates the highlighter data for the given line.
  * Note, that lp might be null, in which case all highlight data is discarded.
  */
 void highlight_invalidate(HIGHLIGHTER* pHighlighter, LINE* lp) {
 	pHighlighter->h_lastLine = -1;
-	pHighlighter->h_lastLinePointer = 0;
-	for (int i = 0; i < WINDOW_SIZE; i++) {
-		if (lp == pHighlighter->h_lines[i].s_linePointer) {
-			pHighlighter->h_lines[i].s_lineEndType = UNKNOWN;
-			break;
-		}
+	if (lp == pHighlighter->h_lastLinePointer) {
+		pHighlighter->h_lastLinePointer = 0;
 	}
 }
 
@@ -185,17 +180,11 @@ unsigned char* highlight_calculate(HIGHLIGHTER* pData, FTABLE* fp, LINE* lp, lon
 /*
  * Return a syntax highlighter for a given grammar;
  */
-HIGHLIGHTER* highlight_getHighlighter(GRAMMAR* pszGrammar) {
+HIGHLIGHTER* highlight_getHighlighter(GRAMMAR* pGrammar) {
 	HIGHLIGHTER* pResult = calloc(1, sizeof(HIGHLIGHTER));
-	pResult->h_calculate = highlight_demo;
-	pResult->h_tokenTypeToStyleTable[SL_COMMENT] = FS_COMMENT;
-	pResult->h_tokenTypeToStyleTable[ML_COMMENT] = FS_COMMENT;
-	pResult->h_tokenTypeToStyleTable[OPERATOR] = FS_OPERATOR;
-	pResult->h_tokenTypeToStyleTable[STRING] = FS_STRING;
-	pResult->h_tokenTypeToStyleTable[IDENTIFIER] = FS_IDENTIFIER;
-	pResult->h_tokenTypeToStyleTable[KEYWORD] = FS_KEYWORD;
-	pResult->h_tokenTypeToStyleTable[OTHER] = FS_NORMAL;
-	pResult->h_tokenTypeToStyleTable[UNKNOWN] = FS_NORMAL;
+	pResult->h_grammar = pGrammar;
+	pResult->h_calculate = highlight_usingGrammar;
+	grammar_initTokenTypeToStyleTable(pGrammar, pResult->h_tokenTypeToStyleTable);
 	highlight_invalidate(pResult, 0);
 	return pResult;
 }