+++ include/grammar.h
@@ -17,34 +17,33 @@
 
 #ifndef GRAMMAR_H
 
-typedef struct tagGRAMMAR_PATTERN {
-	struct tagGRAMMAR_PATTERN* next;	// linked list of grammar patterns.
-	char  name[32];						// The name of the match - should use hierarchical names as for example 'comment.line' or 'keyword.control.java'
-										// The names may be "matched" against style selectors specified below
-	char* match;						// a simple RE pattern to match this pattern
-	char  begin[12];					// mutually exclusive with match, one may define a begin and end marker to match. May span multiple lines
-	char  end[12];						// the end marker maybe e.g. '$' to match the end of line.
-} GRAMMAR_PATTERN;
-
-typedef struct tagSTYLE_SELECTOR {
-	struct tagSTYLE_SELECTOR* next;		// linked list of style selectors.
-	char styleName[32];					// the name of a registered editor style - note, the maximum length of style names in PKS-EDIT is 32
-	char patternName[32];				// Pattern name or prefix of a pattern name like 'keyword' to match 'keyword.control' and 'keyword.operator'
-} STYLE_SELECTOR;
-
 /*
  * A grammar describes the grammar of a "language" of an input file editable by PKS Edit
  * for purpose like syntax highlighting or section folding etc. Grammars are loaded from
  * files named [scopeName].grammar.json located in the PKS_SYS directory.
  */
-typedef struct tagGRAMMAR {
-	struct tagGRAMMAR* next;			// Grammars are maintained - like many things in PKS Edit - in a linked list.
-	char scopeName[32];					// Unique name of the grammar. One can associate document types with the scope name of a grammar
-										// in the document type definition.
-	char* folderStartMarker;			// regular expression to define the start of foldable regions.
-	char* folderEndMarker;				// regular expression to define the end of foldable regions.
-	GRAMMAR_PATTERN* patterns;			// The patterns defined for this grammar.
-} GRAMMAR;
+typedef struct tagGRAMMAR GRAMMAR;
+
+typedef enum {
+	UNKNOWN = 0,
+	INITIAL,
+	CUSTOM_STATE
+} LEXICAL_STATE;
+
+#define MAX_LEXICAL_ELEMENT 100
+#define MAX_TOKEN_TYPE	64
+
+typedef struct tagLEXICAL_ELEMENT {
+	LEXICAL_STATE le_state;			// the lexical state determined for the next le_length characaters.
+	int le_length;					// the length of the lexical state.
+} LEXICAL_ELEMENT;
+
+/*
+ * This method will parse the characters in pszBuf (assumed length is lLength) and determine the lexial
+ * tokens defined as LEXICAL_ELEMENTs to be placed in the result buffer. If successful return the number of
+ * lexical elements actually found.
+ */
+extern int grammar_parse(GRAMMAR* pGrammar, LEXICAL_ELEMENT pResult[MAX_LEXICAL_ELEMENT], LEXICAL_STATE starState, char* pszBuf, size_t lLength);
 
 /*
  * Destroy all loaded grammars and style selectors.
@@ -56,6 +55,11 @@ extern void grammar_destroyAll();
  * on the fly.
  */
 extern GRAMMAR* grammar_findNamed(char* pszGrammarName);
+
+/*
+ * Initialize the style translation table for a given grammar.
+ */
+extern void grammar_initTokenTypeToStyleTable(GRAMMAR* pGrammar, unsigned char tokenTypeToStyleTable[MAX_TOKEN_TYPE]);
 
 #define GRAMMAR_H
 #endif